New functions
Added some new things to the game:
- Loot system working perfectly fine
- Character panel working  fine
- can equip/dequip items
- manage to mage a regen potion
- chest that works perfectly fine you can store items closed and when you check again they will stay there yo ucan also use multiple chests.
-Tooltip (it will describe an object)
Finally I added a "Hand" to drag items over the skill bar and they are kind of functional still can't unequip them from the bar.

+++ The legacy of Kresnik/Assets/Scripts/GUI/HandScript.cs
@@ -34,7 +34,11 @@ public class HandScript : MonoBehaviour
     void Update()
     {
         icon.transform.position = Input.mousePosition;
-        DeleteItem();
+
+        if (Input.GetMouseButton(0) && !EventSystem.current.IsPointerOverGameObject() && MyInstance.MyMoveable != null)
+        {
+            DeleteItem();
+        }
     }
 
     public void TakeMoveable (IMoveable moveable)
@@ -57,19 +61,26 @@ public class HandScript : MonoBehaviour
     {
         MyMoveable = null;
         icon.color = new Color (0, 0, 0, 0);
+        InventoryScript.MyInstance.FromSlot = null;
     }
 
-    private void DeleteItem()
+    public void DeleteItem()
     {
-        if (Input.GetMouseButton(0) && !EventSystem.current.IsPointerOverGameObject() && MyInstance.MyMoveable != null)
+        if (MyMoveable is Item)
         {
-            if (MyMoveable is Item && InventoryScript.MyInstance.FromSlot != null)
+            Item item = (Item)MyMoveable;
+            if (item.MySlot != null)
             {
-                (MyMoveable as Item).MySlot.Clear();
+                item.MySlot.Clear();
             }
-
-            Drop();
-            InventoryScript.MyInstance.FromSlot = null;
+            
+            else if (item.MyCharacterButton != null)
+            {
+                item.MyCharacterButton.DequipArmor();
+            }
         }
+
+        Drop();
+        InventoryScript.MyInstance.FromSlot = null;
     }
 }
